requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
anyio=3.3.0=py38h578d9bd_0
argon2-cffi=20.1.0=py38h497a2fe_2
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=3.3.1=pyhd8ed1ab_0
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.14.6=py38ha65f79e_0
chardet=4.0.0=py38h578d9bd_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
click=7.1.2=pyh9f0ad1d_0
configparser=5.2.0=pyhd8ed1ab_0
cryptography=3.4.7=py38ha5dfef3_0
cudatoolkit=11.1.74=h6bb024c_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.4.1=py38h709712a_0
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
docker-pycreds=0.4.0=py_0
entrypoints=0.3=pyhd8ed1ab_1003
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.6.3=py38h578d9bd_0
intel-openmp=2021.2.0=h06a4308_610
ipykernel=6.0.3=py38hd0cf306_0
ipython=7.25.0=py38hd0cf306_1
ipython_genutils=0.2.0=py_1
jedi=0.18.0=py38h578d9bd_2
jinja2=3.0.1=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py38h578d9bd_0
jupyter_server=1.10.1=pyhd8ed1ab_0
jupyterlab=3.1.1=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.6.1=pyhd8ed1ab_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.1=h27cfd23_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.15.8=h780b84a_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h2531618_0
markupsafe=2.0.1=py38h497a2fe_0
matplotlib-inline=0.1.2=pyhd8ed1ab_2
mistune=0.8.4=py38h497a2fe_1004
mkl=2021.4.0=h06a4308_640
mkl-service=2.3.0=py38h27cfd23_1
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.1=py38ha9443f7_2
nbclassic=0.3.1=pyhd8ed1ab_1
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.1.0=py38h578d9bd_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
notebook=6.4.0=pyha770c72_0
numpy=1.20.2=py38h2d18471_0
numpy-base=1.20.2=py38hfae3a4d_0
olefile=0.46=py_0
opencv-python=4.5.5.62=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1m=h7f8727e_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.1.3=py38he6710b0_0
pandoc=2.14.1=h7f98852_0
pandocfilters=1.4.2=py_1
parso=0.8.2=pyhd8ed1ab_0
pathtools=0.1.2=py_1
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.0.0=py38h9a89aac_0
pip=21.1.3=py38h06a4308_0
prometheus_client=0.11.0=pyhd8ed1ab_0
promise=2.3=py38h578d9bd_5
prompt-toolkit=3.0.19=pyha770c72_0
protobuf=3.15.8=py38h709712a_0
psutil=5.8.0=py38h27cfd23_1
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.9.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.17.3=py38h497a2fe_2
pysocks=1.7.1=py38h578d9bd_3
python=3.8.10=h12debd9_8
python-dateutil=2.8.2=pyhd8ed1ab_0
python-lmdb=0.99=py38h709712a_0
python_abi=3.8=2_cp38
pytorch=1.9.0=py3.8_cuda11.1_cudnn8.0.5_0
pytz=2021.1=pyhd8ed1ab_0
pyyaml=5.4.1=py38h497a2fe_0
pyzmq=22.1.0=py38h2035c66_0
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd8ed1ab_0
requests-unixsocket=0.2.0=py_0
scipy=1.7.3=py38hc147768_0
send2trash=1.7.1=pyhd8ed1ab_0
sentry-sdk=1.5.6=pyhd8ed1ab_0
setuptools=52.0.0=py38h06a4308_0
shortuuid=1.0.8=py38h578d9bd_0
six=1.16.0=pyhd3eb1b0_0
smmap=3.0.5=pyh44b312d_0
sniffio=1.2.0=py38h578d9bd_1
sqlite=3.36.0=hc218d9a_0
subprocess32=3.5.4=py_1
termcolor=1.1.0=py_2
terminado=0.10.1=py38h578d9bd_0
testpath=0.5.0=pyhd8ed1ab_0
tk=8.6.10=hbc83047_0
torchaudio=0.9.0=py38
torchvision=0.10.0=py38_cu111
tornado=6.1=py38h497a2fe_1
tqdm=4.61.1=pyhd8ed1ab_0
traitlets=5.0.5=py_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd8ed1ab_0
wandb=0.12.10=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=0.57.0=py38h578d9bd_4
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
yaspin=2.1.0=pyhd8ed1ab_0
zeromq=4.3.4=h9c3ff4c_0
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
